+++ solution/WellFired.Guacamole.Unity.Editor/Extensions/UnityVector2Extensions.cs
@@ -5,19 +5,19 @@ using WellFired.Guacamole.Data.Annotations;
 namespace WellFired.Guacamole.Unity.Editor.Extensions
 {
 	public static class UnityVector2Extensions
-	{
+	{	
 		// ReSharper disable once InconsistentNaming
 		[PublicAPI]
 		public static UISize ToUISize(this Vector2 source)
 		{
-			return UISize.Of((int) source.x, (int) source.y);
+			return UISize.Of(source.x, source.y);
 		}
 
 		// ReSharper disable once InconsistentNaming
 		[PublicAPI]
 		public static UILocation ToUILocation(this Vector2 source)
 		{
-			return UILocation.Of((int) source.x, (int) source.y);
+			return UILocation.Of(source.x,  source.y);
 		}
 	}
 }
