+++ solution/WellFired.Guacamole/Layouts/ILayoutable.cs
@@ -4,8 +4,8 @@ namespace WellFired.Guacamole.Layouts
 {
     public interface ILayoutable
     {
-        int X { get; set; }
-        int Y { get; set; }
+        float X { get; set; }
+        float Y { get; set; }
         UIRect RectRequest { get; set; }
         UIRect ContentRectRequest{ get; set; }
         LayoutOptions HorizontalLayout { get; set; }
